requirements.txt
Django==3.2.9
django-filter==21.1
djangorestframework==3.12.4
Pillow==8.4.0
python-decouple==3.5

